+++ Codebase/Trinity Core MK3/src/LedManager/ledManager.h
@@ -23,9 +23,9 @@ public:
   void setBrightness(uint8_t brightness);
   uint8_t getPanelAmount();
   uint8_t getPanelDiodeAmount(uint8_t panelNumber);
-  void setPanelData(uint8_t panelNumber, uint8_t brightness, uint8_t effect, uint8_t colour, uint8_t offset, uint8_t speed, bool repeat, bool detailed);
+  void setPanelData(uint8_t panelNumber, uint8_t brightness, uint8_t effect, uint8_t colour, uint16_t offset, uint8_t speed, bool repeat, bool detailed);
   void setPanelCustomData(uint8_t panelNumber, uint8_t customRGBAmount, ColourRGB *customRGB[AMOUNTOFCOLOURS]);
-  void setPanelDiodeData(uint8_t panelNumber, uint8_t diodeNumber, uint8_t brightness, uint8_t effect, uint8_t colour, uint8_t offset, uint8_t speed, bool repeat);
+  void setPanelDiodeData(uint8_t panelNumber, uint8_t diodeNumber, uint8_t brightness, uint8_t effect, uint8_t colour, uint16_t offset, uint8_t speed, bool repeat);
   void setEnabled(bool enabled);
   String convertToTansmission();
   String convertPanelToTransmission(uint8_t panelNumber);
